LayoutTests/platform/blackberry/accessibility/aria-list-and-listitem-expected.txt



This tests that the ARIA roles of list and listitem map correctly to Mac accessibility roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL obj.childAtIndex(0).role should be AXRole: AXList. Was .
FAIL obj.childAtIndex(0).childAtIndex(0).role should be AXRole: AXGroup. Was .
FAIL obj.childAtIndex(0).childAtIndex(1).role should be AXRole: AXGroup. Was .
PASS successfullyParsed is true

TEST COMPLETE
